% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{rosepine_palette}
\alias{rosepine_palette}
\alias{rose_pine}
\alias{rose_pine_moon}
\alias{rose_pine_dawn}
\title{Rosé Pine official color palette(s)}
\format{
An object of class \code{list} of length 15.

An object of class \code{list} of length 15.

An object of class \code{list} of length 15.
}
\source{
https://rosepinetheme.com/
}
\usage{
rose_pine

rose_pine_moon

rose_pine_dawn
}
\description{
These are the official colors as listed on
the \href{https://rosepinetheme.com/palette/}{Rosé Pine palette page}
in August 2024.
}
\keyword{datasets}
