% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{rosepine_rstheme}
\alias{rosepine_rstheme}
\title{Create Rosé Pine .rstheme file}
\usage{
rosepine_rstheme(variant = NULL, install = TRUE, ...)
}
\arguments{
\item{variant}{A string - A string defining the theme palette variant to use
\href{https://rosepinetheme.com/palette/}{Rosé Pine Variants}}

\item{install}{A Boolean - generate and install/apply the .rstheme file
(TRUE), or generate the .scss file without installing/applying}

\item{...}{}
}
\description{
Create Rosé Pine .rstheme file
}
